Fix build after upstream changes in postgres 9.5 -- 12 made functions AcquireDeletionLock() and ReleaseDeletionLock() extern rather than static in:
   https://github.com/postgres/postgres/commit/5be594caf818e0b5e33f8dec191f2707394a6d95

The function definitions in the postgresql sources are identical to
those in the pglogical sources, so let's just link against the
postgresql versions.

postgresql-9.4 is EoL, so didn't receive the same changes in the
latest round of postgresql updates.

While here, fix another of the patches which should no be applied when
compiling against all versions of postgresql.  Plus fix the pkg-plist
-- how did this port ever package previously?