MFH: r535706
Mark as BROKEN on powerpc64:

  include/config_distro_bootcmd.h:279:3: sandbox EFI support is only supported on ARM and x86

This is even after adding the fix for 12-STABLE of compiler:c11 to USES,
which at least allows the build to get that far.

I took a look at this one time, thinking it would be easy, but I was not
able to come up with a quick answer.

Approved by:	portmgr (tier-2 blanket)