Pre Qt 5.15 housekeeping: textproc/kreport
This is probably a conflagration of CMake, Qt and other updates all
happening at once. In poudriere, build of databases/kexi fails with

ld: error: undefined symbol: KReportGroupTracker::qt_metacast(char const*)
>>> referenced by mocs_compilation.cpp
>>>               src/plugins/reports/CMakeFiles/kexi_reportplugin.dir/kexi_reportplugin_autogen/mocs_compilation.cpp.o:(KRScriptFunctions::qt_metacast(char const*))

The problem lies in textproc/kreport, which has a QObject-derived
class with no out-of-line virtual methods, and no clear path
to have moc called on it. This patch splits out the virtual
destructor, adds a .cpp file containing it, and causes automoc
to run normally on the class. This makes all the QObject machinery
appear in the shared library, and kexi can use it.