MFH: r541061
net/rpki-client: Fix creaton of cache and db dir

Fix the creation of the cahce and db dir for rpki-client.  When running the
build as root, the rpki-client build infra tries to change the owner of the
cache and db dir, which fails since the user isn't created until later.
Since we deal with special perimissions and owners/groups of files in
pkg-plist anyway, there is no need to try to change the owner/group from the
rpki-client build infra, so add a patch that just creates the directories
without changing ownership.

This fixes the build of rpki-client on the package cluster.

Approved by:	ports-secteam (joenum)