MFH: r543974
x11/dmenu: Get the latest bug fixes from upstream

Users are still reporting issues with the dmenu version 4.9. The patch we
have currently in the ports tree was meant to be a temporary fix as we
waited for a new dmenu release. A new release is not coming apparently,
so let's just get the latest patches from upstream.

Also, switch to a GitHub-hosted mirror of the dmenu repository
as it is painful to get artifacts from https://git.suckless.org/
without a Git client.

While here, replace REINPLACE_CMD with patches.

Reported by:	Scott Robbins
Approved by:	ports-secteam (joneum)