Update to 1.66
As of this release, the package now updates files which were previously
updated by the app.  In keeping with not having modified files within
WWWDIR, the following files are now supplied by the package and are
created at the same time the vendor tarball is manually created.

The files in question are located in

* bootstrap/cache
* html/js/lang

These files should now be static and not modified by the application,
which would create them if not found. These files will get updated
with each new release.