security/gnupg1: Build with -fcommon
With llvm 11, builds are using -fno-common by default.
I've tried to fix gnupg1 to build with this, but it is qyite hard and
requires a lot of patches.  In the meantime, upstream code has changed
significantly, and gnupg1 is old, so just switch it to use -fcommon instead,
to make it build.

MFH:		2020Q3 (implicit, -fno-common fixes, ok by joenum)