Fix the build with clang 11.
This was fixed upstream as part of a larger commit on May 6th 2017,
and dim@ now encountered the same and fixed the specific issue for
our GCC 7 port.  Newer versions should not exhibit this.

(See the new files/patch-gcc_config_i386_i386.c for more details.)

PR:		248755
Submitted by:	dim