MFH: r546029
security/bsmtrace: backport the -fno-common fix

This port is the older version of bsmtrace and should go away in due time,
but it's easy enough to backport the fix in the meantime. This version did
need an additional fix to log.{c,h} that was no longer needed in 3.x.

PR:		248756
Approved by:	csjp (maintainer, also discussed with)

Approved by:	ports-secteam (implicit, -fno-common build fix)