Mk/Use/python.mk: Do not surround package list entries in quotes
When python package lists are generated by distutils, any file entry that
includes a space will be surrounded in double quotes.  This is unnecessary
and causes problems elsewhere such as when stripping ${PREFIX} for staging
or checking the package list in stage-qa.

PR:		248981
Approved by:	kevans, portmgr (mat), python (koobs)
Differential Revision:	https://reviews.freebsd.org/D26221