graphics/drm-devel-kmod: Update to v5.4.62
This update the drm drivers to be in sync with Linux v5.4.62
This also set the default for psr (Panel Self Refresh) for i915 to disable.
There is some issue in our code that make the console not refreshing as it
should and since I don't have the hardware to reproduce simply change the default
for us.
This also change the power_well default in i915 to have them always enabled. We likely
miss something related to this for FreeBSD and this cause a spam on the console when
using HDMI.

Sponsored by:	The FreeBSD Foundation