Return image based on Accept headers rather than URL
In my setup, I find that a lot of the time the check that's currently in place doesn't hold true which results in me seeing blank pages when I click a link with a tracker behind it rather than the page indicating a block.

While I know this is a minor inconvenience, it probably makes sense to use the hints provided by the browser about the kind of data it's expecting rather than the URL - In this case, if the url is fetched via, e.g. an `<img>` tag, all browsers I have tested will make their top priority an image mime in the Accept header, which will trigger this new check.

Paired with the SSL improvements in https://github.com/pfsense/FreeBSD-ports/pull/778 I think this could make for much better UX.

I'm also considering methods for making it possible to "proceed" in cases where the blocked domain is just serving as a redirect, if that would be interesting.