Xastir 2.1.6 was released in April of 2020.
FreeBSD's port still has 2.1.4 (over a year old), and that port was
recently patched to fix the build with GCC 10 by adding "-fcommon"
to the CFLAGS.  Xastir 2.1.6 has had its GCC 10 compatibility fixed
so that sort of hack is no longer needed in the port if it were
updated to the new version.

There are also numerous other improvements over 2.1.4, including removal
of obsolete options, and consolidation of some options that had
previously been set up as independently selectable, but were in fact
completely dependent on each other.  The build is simpler now,
as some of these things that were previously user-selectable are
automatic now -- notably, if shapelib support is requested and
shapelib is found, pcre is also required *and* automatically enables
dbfawk. The code that previously allowed shapelib support without
dbfawk (and therefore pcre) has been removed.

This means that the current Makefile, which has no option for deselecting
shapelib, but does have an option for selecting or deselecting DBFAWK,
needs a little changing:  remove DBFAWK as an option and make pcre
a required package along with shapelib (instead of dependent on
the DBFAWK setting) and all should be well.

Submitted by:	russo@bogodyn.org
Approved by:	Maintainer