Switch MASTER_SITES to a repository, which contains FreeBSD patches
I've created a repository, which incorporates our patches into the build
system of Rarian so that we could use CONFIGURE_ARGS instead of patches and
REINPLACE_CMD's.

Also, while here, drop the patch replacing LINK with CXXLINK in the build
system of the port. It's just not needed anymore.

I'm leaving the patch modifying --version behavior as it is. It's probably
a bad idea that we decided to modify upstream behavior in the first place
but at this point it probably makes little sense to change the behavior
after so many years.