devel/lua-luacheck: update to just before the unreleased 0.25.0
The intervening changes are many, with new checks and a switch to luarocks
as upstream. This intermediate version also contains some Lua 5.4 support.
Note that Lua 5.4 attributes are currently parsed, but ignored.

argparse is no longer bundled with luacheck, so we'll grab argparse from
devel/lua-argparse.

The build system in luacheck got removed in favor of just using luarocks.
This is currently incompatible with how ports wants to work, so I've changed
it to the simpler approach that was previously advised to just copy the src/
files over and now installing a script from FILESDIR that matches roughly
what we were ending up with before to get the correct package.path.

Miscellaneous indentation fixes included.