net/iplog: DEPRECATE Unmaintained and dead upstream
This port only build in 12 because for some reason uses clang 8 to build. It
does not build in 11 or 13-CURRENT. Even in 12 it shows tons of dangerous
warnigns due to bad formed castings[1].

The project is also abandoned upstream so we better mark this for removal.

[1] https://docs.freebsd.org/cgi/getmsg.cgi?fetch=46148465+0+current/freebsd-pkg-fallout

Submitted by:   udo@schweigert-muc.de (maintainer)
MFH:	2020Q4