new port: devel/easyexif
EasyEXIF is a tiny, lightweight C++ library that parses basic information out of
EXIF files. It uses only the std::string library and is otherwise pure C++.

It is made of one .h file and one .cpp file.

This is a new dependency for the upcoming update of devel/vcglib[1]

[1] https://github.com/cnr-isti-vclab/vcglib/commit/dd8c26474dfa3ee2cde2e4c39366fe9df41e667d