Port of	the FreeBSD calendar program from -CURRENT
This version of	the calendar program includes data file locations
populated by the calendar-data port in the search path.

It is meant to provide this feature to users of	FreeBSD	releases that
have not received the updated calendar program,	yet.

Development of the calendar program will	continue in the base system
and the	port will be updated to	reflect	changes	made to	the calendar
program	in FreeBSD-CURRENT.

Different from the version in -CURRENT this port does not warn about
missing calendar files provided by the calendar-data port since it has
this port declared as a run-time dependency and it can therefore be
assumed that the data port has been installed with this port.

This port uses a preliminary location for the repository to fetch from.

It is expected to be updated to an official repository with updated sources
that eliminate the need for the patches that are currently required.