u-boot-sopine-spi: Add new port
While the u-boot-sopine port can boot from SPI on the Pine64-LTS and Sopine
module the env is saved and read from the eMMC.
This is a modified sopine u-boot that only read and store the env in the spi
flash available on thoses boards.
Additionally there is also a sopine-spi-flasher.img that can be burnt on a sdcard
and when booted it will program the spi flash.