[NEW] sysutils/gitwatch: Bash script to watch a file or folder and commit changes to a git repo
A bash script to watch a file or folder and commit changes to a git repo
What to use it for?

* config files: some programs auto-write their config files, without waiting
for you to click an 'Apply' button; or even if there is such a button, most
programs offer you no way of going back to an earlier version of your settings.
If you commit your config file(s) to a git repo, you can track changes and go
back to older versions. This script makes it convenient, to have all changes
recorded automatically.
* document files: if you use an editor that does not have built-in git support
(or maybe if you don't like the git support it has), you can use gitwatch to
automatically commit your files when you save them, or combine it with the
editor's auto-save feature to fully automatically and regularly track your
changes.

WWW: https://github.com/gitwatch/gitwatch

Sponsored by:	Bounce Experts