Fix slow loading of HAProxy widget on dashboard, buy avoiding reapeated costly permissions checks.
In our current setup we have a few docen frontends with a few backends under the hood. So for each server this is calling userHasPrivilage, which loads an parses config file everytime.. adding ~2s to each loop iteration. This patch changes the code so privilegies are checked just once (before entering the foreach loop).