textproc/bsdgrep: fix the build on recent -CURRENT
A seatbelt was added to throw a warning if a "SUBDIR." variable appeared
non-empty, since that indicates something was trying to expand a MK_
variable most likely and either opts.mk wasn't included or the name of the
variable used is wrong.

In this particular case, we're repackaging a part from base. I had stripped
the src.opts.mk include but hadn't thought to replace it with bsd.opts.mk
for userlandy stuff.

This also goes as far as fixing the test build, but doesn't yet hook up the
tests. On FreeBSD <13 these need a kyua/atf test dep, but I'd like to do a
sweep and see if there are any other ports (e.g. from NetBSD base or FreeBSD
base) that also have a dependency to weigh in on if this should just be a
one-off osversion check or a kyua/atf USES.

MFH:		2020Q4 (blanket: build fix)