emulators/qemu-user-static: update to 6f1319e1138c
Highlights:
- Fix stack allocation to behave more like FreeBSD really does[*].
- Raised the stack size on ARM, which should solve some limited cases of
  mysterious segfaults (e.g. lang/gcc9).
- Provide hw.ncpu and AT_NCPUS based on cpu affinity to more effectively
  limit cpu usage under qemu-user-static.

[*] We start out with a limited initial stack size, then login(1) raises
the stack size up to the maximum possible for the system (arch-based).
qemu-user-static will now calculate a maximum stack size based on the
host's configured rlimit and the qemu-bsd-user specified max for the arch
in question, then immediately raise the stack allocation size to the
maximum. This is especially important because we don't actually handle
stack growth.

post-patch is also dropped entirely, as these sed invocations have done
nothing since versions of qemu back in ~2007, AFAICT. CFLAGS/LDFLAGS get
sucked in during configure and qemu's built with -fno-strict-aliasing as
a default for even longer.