devel/grex: add new port
grex is a command-line utility that is meant to simplify the often
complicated and tedious task of creating regular expressions. It
does so by automatically generating a single regular expression from
user-provided test cases. The resulting expression is guaranteed to
match the test cases which it was generated from.

WWW: https://github.com/pemistahl/grex/

PR:		251424
Submitted by:	Petteri Valkonen