Add py-pygeos 0.8
PyGEOS is a C/Python library with vectorized geometry functions. The geometry
operations are done in the open-source geometry library GEOS. PyGEOS wraps these
operations in NumPy ufuncs providing a performance improvement when operating on
arrays of geometries.

A universal function (or ufunc for short) is a function that operates on
n-dimensional arrays in an element-by-element fashion, supporting array
broadcasting. The for-loops that are involved are fully implemented in C
diminishing the overhead of the Python interpreter.

WWW: https://github.com/pygeos/pygeos