net/openmpi: Add an option to enable AVX support
OpenMPI 4.1 fails to detect if the host supports AVX instructions and will fail to build if it doesn't [1].

Also, include the ABI version to the hwloc2 library dependency. If the user has devel/hwloc (and not devel/hwloc2) installed it will satify the dependency check anyway and link against the wrong lib.

[1] - https://github.com/open-mpi/ompi/issues/8306

PR:		252506