net/chrony-lite: Slimmed down chrony without NTS support
chrony is a pair of programs which are used to maintain the accuracy of the
system clock on a computer; the two programs are called chronyd and chronyc.

chronyd is a daemon which runs in background on the system.  It obtains
measurements via the network of the system clock's offset relative to time
servers on other systems and adjusts the system time accordingly.  For
isolated systems, the user can periodically enter the correct time by hand
(using chronyc).  In either case, chronyd determines the rate at which the
computer gains or loses time, and compensates for this.  chronyd implements
the NTP protocol and can act as either a client or a server.

chronyc provides a user interface to chronyd for monitoring its performance
and configuring various settings.  It can do so while running on the same
computer as the chronyd instance it is controlling or a different computer.

WWW: http://chrony.tuxfamily.org/

PR:		252584
Submitted by:	myself