devel/serd: Patch for aligned_alloc removal
The aligned_alloc is removed in the upstream in the c4c4ec51 commit. It creates
problems for MacOS and FreeBSD, apparently. This patch is recreating that commit
using makepatch, as latest version (0.30.8) and master diverged too much to
apply just that commit (I tried, and failed with PATCHFILES).

PR:	252859
Submitted by:	meka@tilda.center (maintainer)