Updated Seafile/Seahub to 8.0.3
Changed to -server branch, some errors were creeping over
from the pro version not being compatible with the community
edition. Previously this was never an issue as you must manually
upgrade to the pro version for the new features.

Disabled seafile-controller config checking prior to startup.
The check only ran on ccnet previously and this was changed to
seafile which never implemented the function.
https://forum.seafile.com/t/seafile-controller-t-functionality-broken-on-8-x/13738

This version adds compatible with Django 2.2, this will require
coordination with multiple port devs to migrate this over
properly.

Changes: https://manual.seafile.com/changelog/server-changelog/

PR:		252715