Add mesa-dri dependency for qt5-gui
There are cases (headless, no xorg-server) where a qt5-gui application
still needs to run. It can use swrast, but only if that is installed.
Add the mesa-dri dependency to make that happen automatically.

For "normal" Qt5 usage, that is, Qt5 on the desktop, xorg-server already
pulls this in. The only change is for systems with no X11, no Wayland,
but with Qt5 GUI applications.

PR:		241490
Submitted by:	madpilot
Reported by:	madpilot