Fix Qt5 ports
Thanks to antoine@ for figuring out that ${REINPLACE_CMD} does not
handle '\t' in all supported versions -- so replacing the tabs in
the script kept the tabs, but called /usr/bin/ca and /usr/bin/sor
instead. Switched to indenting with spaces instead, so and only
strip leading spaces at that, so the scripts can be slightly
easier to indent / layout. Add comment pointing out how to
easily read the shellscript among all the minifying-magic.

danfe@ pointed to a naming mix-up; because the variables were
used consistently, this had no effect but a weirdly-named temporary
file in ${WRKSRC}. Swapped the names the right way around.

Bump PORTREVISION on qt5-core again to force another round of
rebuilds.

Reported by:	danfe, antoine