Improve FreeBSD packaging in devel/cmake
The OPTION CPACK in CMake enables building FreeBSD packages directly
from cmake / cpack. This can be useful, for instance, with not-yet-
packaged software, or building packages for proprietary software.
This patch has been submitted upstream; it resolves a disagreement
between cmake and libpkg about the name of the resulting package.