pfSense-upgrade: Send notice about WireGuard
We need to notify users they can't proceed with upgrade while there are
enabled WireGuard tunnels on current config but GUI is not able to catch
the error message printed by pfSense-upgrade.

To workaround this issue on version GUI is not fixed yet I've created a
function to send a notice.  Also added the code to send this notice when
pfSense-upgrade -c is called to make sure users are going to be notified
in advance since this check runs on crontab every day.