net/radvd: Fix failure after SIGHUP
On SIGHUP radvd re-reads its configuration and re-joins the all-routers
multicast group. On FreeBSD this fails because the kernel rejects
attempts to join a group we're already a member of.
Newer kernels will tell us that this is what's happening by returning
EADDRINUSE, so don't consider that to be a failure.

PR:	https://redmine.netgate.com/issues/6287