Fix 404 error to firewall_virtual_ip.php in doco
The current pfsense haproxy 0.61_5 package generates a 404 when clicking on the Virtual IP link.
It links to /haproxy/firewall_virtual_ip.php instead of /firewall_virtual_ip.php as the link is relative to the /haproxy sub-directory.
This makes it look as if the haproxy project is meant to ship with a firewall_virtual_ip.php page, instead of linking to the standard pfsense one under Firewall / Virtual IPs.

(cherry picked from commit d32312de35cecd94a77295f68a7bac9a39b957ce)