php-pfSense: Set the proper label on a rule - Regression #12977
Rather than blindly assign label[0] to the exported rule 'label' value in the
associative array, iterate through all non-empty labels for each rule and build
an additional associative array 'all_labels' keyed on the label prefix. If a
label has no prefix or, preferentially, is prefixed with 'USER_RULE: ', assign
it to the legacy 'label' value.