Fix WWW in parent/child ports
Many of the WWW are overwritten later which means the wrong value
is used. This did not happen before where the children were either
 a) just using the pkg-descr from the parents
 b) or had their own separate pkg-descr with custom WWW

Use WWW?= in parents when the child's WWW is different.

Children that use the same WWW as the parent can just inherit it,
i.e., the child WWW can be removed.

Approved by:	portmgr (implicit)