Eliminate races that caused incomplete alias tables. Fixes #9269
Thread synchronization originally relied on condvars, which allowed a race to
exist if a thread was attempting to signal another that was not at that moment
already waiting on the condvar. This race was most notable when an alias table
was added that referenced known host entries for IP addresses, and the host
thread (without needing to perform a host lookup) would signal its newly created
action thread before it was waiting on the condvar, causing the table associated
with it to not be updated until the next interval that woke up its host thread
or another change was made to the aliases.

To alleviate this, the condvars are replaced with semaphores, which are more
appropriate for the type of synchronization being done. Additional changes are
made related to thread synchronization:

 * The interval for check_hostname() to wake was given an additional interval%30
   seconds for an inexplicable reason. This is removed.
 * Relocated main_lock initialization prior to registration of the signal
   handler, which expects it to be initialized.
 * Removed host thread signal from host_del(), as it is only called as a result
   of merge_config() which signals all host threads anyway after it has
   completed its work. The additional signal just mades the host actions run
   twice for a single config sync.