Fix orphaned threads, SIGHUP deadlocks, and more races. #9269
Make check_action() threads exit upon host_del() by checking for a null hostname
on wakeup and posting to an exit semaphore that host_del() will wait on before
destroying the act structure. Prior to this change, filterdns would accumulate
threads for each host->table relationship on each configuration change, but when
deleting their associated action structs it would just leave them sleeping.

Move exit code to merge_config() so it is not run within another thread, and
add signalling and thread state to the merge_config() thread so it can be told
to exit its loop with a THR_DYING state and a post to the semaphore.

In action_del(), check for null hostname before derefencing it in the log message
for safety.

Remove superfluous main_lock grabbing in check_action() that likely was
introduced as a hack that alleviated some race conditions.